allow hex strings to contain uppercase digits
Currently, the grammar does not support case insensitive hex.
So, #fff parses while #FFF does not. This is a design feature of
atomizer.  However, we can't generate meaningful errors without
parsing them.  This PR extends the grammar to parse these colors
and then checks for uppercase hex colors and generates a warning.